Refactor on user table so that data, styles and the component itself are on separated files
