+++ src/routes/users/table/Styles.js
@@ -1,0 +1,16 @@
+import {makeStyles} from '@material-ui/core/styles';
+
+export const useStyles = makeStyles({
+  centered: {
+    position: 'fixed',
+    left: '55%',
+    top: '45%',
+    color: '#1F271B',
+  },
+  table: {
+    position: 'relative',
+    left: '108px',
+    top: '90px',
+    width: 1100,
+  },
+});
